Add context to leaks
The context is a new option that allow gitleaks to return a fix
number of characters before and after the match/leak. (If there
is multiple lines, the first and last lines are trimmed for
readability)
The context of a leak is filled with a buffer that retains the last
characters read in the current Chunk. (So it is not cross-Chunk)
We will avoid very large line, and also have a bit of context of a
leak.