Add option to scan additions only
Additions are sometimes the only modifications that we need to scan.
All deletions are additions of the past, and in a case where we're
only scanning one line (without context). It is not relevant to have
two times the exact same "leak".
Depending of the needs, if we only want the ones that commited the
leak and not deleted it.
In these cases, the scans would be accelerated :)