A first attempt adding hooks.slack.com/triggers/
As described in issue 1791, the hooks.slack.com/triggers should perhaps
also be detected.

Reference
https://slack.com/help/articles/360041352714-Build-a-workflow--Create-a-workflow-that-starts-outside-of-Slack

Feel free to use, change any part of this proposed fix, or to simply
discard it.

(ps. I did a minor typo in a comment in issue 1791, proposing this
change but that was corrected here. ds.)