fix(pre-commit): Use git-agnostic file-based checking
This fixes gitleaks invocations via pre-commit when there are no changed
to-be-committed (i.e., via `pre-commit run --files` or `pre-commit run
--all-files`). Currently, these will pass even if there would be
gitleaks findings because `gitleaks git` checks the diff only.

When called via pre-commit framework, gitleaks does not need to
git-aware because the framework already detects which files need to be
checked (either the diff when a commit range or changes that are
about-to-be-committed are checked, or the files passed to the `--files`
command line argument).

Fixes https://github.com/gitleaks/gitleaks/issues/1409.