Allow multiple sources for `dir` command
My use for this is for feeding gitleaks an explicit list of changed
files for continuous scanning of pull requests to our repository.

This has been requested in #1888, #1727, #1234, and possibly others.

This has previously been implemented in #1238 but for a much older
version and it seemed easier to re-do this rather than to rebase

There are some "source"-related configuration processing tasks done
in `initConfig()` and `Detector()`; for simplicity, I chose to
require explicit configuration of those items (rather than guessing
per source, having to distinguish files vs dirs, handling different
configuration and ignore files etc.)

I also didn't really understand the logic about figuring out the
relative baseline path with respect to the source, and how should
that work when multiple sources may be present. It seemed easier
to simply disable the baseline in a "multiple source" scenario.